fix(proto): nest fks-proto modules + align distributed schema to call sites

+++ services/forward/src/regime_bridge_server.rs
@@ -774,7 +774,7 @@ impl RegimeBridgeService for RegimeBridgeServer {
                     let seq = inner_for_stream
                         .stream_sequence
                         .fetch_add(1, Ordering::Relaxed);
-                    let mut proto_state = RegimeState::from(&bridged_state).with_sequence(seq);
+                    let mut proto_state = RegimeState::from(&bridged_state).with_sequence(seq as i64);
 
                     // Transition detection
                     let hypo_i32 = proto_state.hypothalamus_regime;
